requirements.txt
albumentations==1.3.0
matplotlib==3.5.1
menpo==0.11.0
mxnet==1.9.1
numpy==1.21.5
opencv_python_headless==4.7.0.72
pandas==1.4.2
Pillow==9.5.0
prettytable==3.6.0
PyYAML==6.0
scikit_image==0.19.2
scikit_learn==1.2.2
scipy==1.7.3
skimage==0.0
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tqdm==4.64.0
